Fix working Hour in nurse if patient Resverd an apoointment and nurse remove it should remove from patient dashboard

+++ app/Http/Controllers/NurseController.php
@@ -168,19 +168,41 @@ class NurseController extends Controller
             {
                 if($from_That_dayTime->diffInDays($to_That_dayTime) == "0"){
                     for($date = clone $from_That_dayTime ; $date->diffInMinutes($to_That_dayTime) >"30";$date->addMinutes(30)){
-                        $res = new Reservation();
-                        $res["reservation At"] = $date;
-                        $res["user_id"] = 0;
-                        $res["Reserved_by_Doctor"] = 1;
-                        $res->save();
+
+                        $check = Reservation::firstWhere("reservation At", $date->toDateTimeString());
+                        if ($check === null) {
+                            $res = new Reservation();
+                            $res["reservation At"] = $date;
+                            $res["user_id"] = 0;
+                            $res["Reserved_by_Doctor"] = 1;
+                            $res->save();
+                        }else{
+                            $check->delete();
+                            $res = new Reservation();
+                            $res["reservation At"] = $date;
+                            $res["user_id"] = 0;
+                            $res["Reserved_by_Doctor"] = 1;
+                            $res->save();
+
+                        }
                     }
                 }else{
                     for($date = clone $x ; $date->diffInMinutes($to_That_dayTime) >"30";$date->addMinutes(30)){
-                        $res = new Reservation();
-                        $res["reservation At"] = $date;
-                        $res["user_id"] = 0;
-                        $res["Reserved_by_Doctor"] = 1;
-                        $res->save();
+                        $check = Reservation::firstWhere("reservation At", $date->toDateTimeString());
+                        if ($check === null) {
+                            $res = new Reservation();
+                            $res["reservation At"] = $date;
+                            $res["user_id"] = 0;
+                            $res["Reserved_by_Doctor"] = 1;
+                            $res->save();
+                        }else{
+                            $check->delete();
+                            $res = new Reservation();
+                            $res["reservation At"] = $date;
+                            $res["user_id"] = 0;
+                            $res["Reserved_by_Doctor"] = 1;
+                            $res->save();
+                        }
                     }
                 }
             }
@@ -213,7 +235,9 @@ class NurseController extends Controller
         $patient_turn['user_id'] = $user->id;
         $patient_turn->save();
         $reservation = Reservation::where("user_id",$user["id"])->whereBetween('reservation At',[Carbon::today()->toDateTime(),Carbon::today()->addHours(22)->toDateTime()])->get();
-        $reservation[0]->delete();
+        if(count($reservation)!=0) {
+            $reservation[0]->delete();
+        }
         return redirect('/nurse');
     }
     public function CheckNotAttend($id){
@@ -222,7 +246,9 @@ class NurseController extends Controller
             $user->patient->Attend = $user->patient->Attend - 1;
         }
         $reservation = Reservation::where("user_id",$user["id"])->whereBetween('reservation At',[Carbon::today()->toDateTime(),Carbon::today()->addHours(22)->toDateTime()])->get();
-        $reservation[0]->delete();
+        if(count($reservation)!=0) {
+            $reservation[0]->delete();
+        }
         return redirect('/nurse');
     }
 
@@ -234,43 +260,49 @@ class NurseController extends Controller
         $dateTimeTo = clone $dateTimeFrom;
         $dateTimeTo->addHours(14);
 
-        $datenow = Carbon::now();
-        $res_check_for_duplicate= Reservation::firstWhere("Reserved_by_Doctor",1);
-        $res_check2_for_duplicate= Reservation::whereBetween('reservation At',[Carbon::today()->toDateTime(),Carbon::today()->addHours(22)->toDateTime()])->where('Reserved_by_Doctor',1)->orderBy('reservation At','DESC')->get();
 
-        if(count($res_check2_for_duplicate)!=0) {
-            $start = clone Carbon::parse($res_check2_for_duplicate[0]["reservation At"]);
-            if($start->isPast()) {
-                for ($date = clone $start; $date->diffInMinutes($datenow) >= "30"; $date->addMinutes(30)) {
-                    $date->addMinutes(30);
+        $datenow = Carbon::now();
+        $start = clone Carbon::today();
+        $start->addHours(8);
+        $today = clone $start;
+        if($start->isPast()) {
+            for ($date = clone $start; $date->diffInMinutes($datenow) > "30"; $date->addMinutes(30)) {
+                $check = Reservation::firstWhere("reservation At", $date->toDateTimeString());
+                if ($check === null) {
                     $res = new Reservation();
                     $res["reservation At"] = $date;
                     $res["user_id"] = 0;
                     $res["Reserved_by_Doctor"] = 1;
                     $res->save();
                 }
+                $today = $date;
             }
-        }elseif(count($res_check2_for_duplicate)==0){
-            $start = clone Carbon::today();
-            $start->addHours(8);
-            if($start->isPast()) {
-                for ($date = clone $start; $date->diffInMinutes($datenow) > "30"; $date->addMinutes(30)) {
-                    $res = new Reservation();
-                    $res["reservation At"] = $date;
-                    $res["user_id"] = 0;
-                    $res["Reserved_by_Doctor"] = 1;
-                    $res->save();
-                }
+            $end = clone $today;
+            $check2 = Reservation::firstWhere("reservation At", $end->toDateTimeString());
+            if($check2 === null ){
+                $res = new Reservation();
+                $res["reservation At"] = $end;
+                $res["user_id"] = 0;
+                $res["Reserved_by_Doctor"] = 1;
+                $res->save();
             }
         }
         $reservedobj = Reservation::whereBetween('reservation At',[$dateTimeFrom->toDateTime(),$dateTimeTo->toDateTime()])->get();
 
         $reserved = array();
+        $isreserved = true;
+
         foreach ($reservedobj as $res){
             $first = Carbon::parse($res["reservation At"])->format('Hi');
             $first = (int)$first;
             array_push($reserved,$first);
+
+            if($res["user_id"] === auth()->user()->id){
+                $isreserved= false;
+            }
         }
+
+        array_push($reserved,$isreserved);
         echo json_encode($reserved);    // Echo Available Appointments Fro Day/Month
     }
